requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.5
django-rest-swagger==2.1.2
djangorestframework==3.7.0
drfdocs==0.0.11
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
openapi-codec==1.3.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
